+++ src/Microsoft.Health.Fhir.Shared.Core/Features/Resources/Delete/ConditionalDeleteResourceHandler.cs
@@ -98,7 +98,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Resources.Delete
         {
             var itemsDeleted = await _deleter.DeleteMultipleAsync(request, cancellationToken);
 
-            return new DeleteResourceResponse(itemsDeleted);
+            return new DeleteResourceResponse((int)itemsDeleted);
         }
     }
 }
